--- apps/zui/src/app/core/hooks/use-auto-select.ts
@@ -1,9 +1,0 @@
-import {RefObject, useLayoutEffect} from "react"
-
-export function useAutoSelect(ref: RefObject<HTMLInputElement>) {
-  useLayoutEffect(() => {
-    if (ref.current) {
-      ref.current.select()
-    }
-  }, [])
-}
