--- apps/zui/src/app/core/hooks/useLayoutUpdate.ts
@@ -1,12 +1,0 @@
-import {useLayoutEffect, useRef} from "react"
-
-export default function useLayoutUpdate(fn, deps) {
-  const first = useRef(true)
-  useLayoutEffect(() => {
-    if (first.current) {
-      first.current = false
-    } else {
-      fn()
-    }
-  }, deps)
-}
