+++ apps/zui/src/ppl/detail/EventTimeline.tsx
@@ -1,4 +1,4 @@
-import useContentRect from "src/app/core/hooks/useContentRect"
+import useContentRect from "src/util/hooks/use-content-rect"
 import {scaleTime, extent} from "d3"
 import React, {memo, useState} from "react"
 import styled from "styled-components"
@@ -6,7 +6,7 @@ import {SecurityEventInterface} from "./models/security-event"
 import EventTag from "./EventTag"
 import {useDispatch} from "src/app/core/state"
 import {viewLogDetail} from "src/js/flows/viewLogDetail"
-import useResizeEffect from "src/app/core/hooks/useResizeEffect"
+import useResizeEffect from "src/util/hooks/use-resize-effect"
 import {isEqual} from "lodash"
 import useCallbackRef from "src/js/components/hooks/useCallbackRef"
 import time from "src/js/models/time"
