--- apps/zui/src/util/hooks/use-ref-check.ts
@@ -1,8 +1,0 @@
-import {useRef} from "react"
-
-export function useRefCheck(value) {
-  const prev = useRef<any>(value)
-  if (value !== prev.current) {
-    prev.current = value
-  }
-}
