+++ apps/zui/src/util/merge-refs.ts
