+++ apps/zui/src/views/preview-load-modal/index.tsx
@@ -10,7 +10,7 @@ import {ResultsControl, useResultsControl} from "./results"
 import {Shaper} from "./shaper"
 import {Sidebar} from "./sidebar"
 import {ResultsGroup} from "./results-group"
-import useSelect from "src/app/core/hooks/use-select"
+import useSelect from "src/util/hooks/use-select"
 import {ErrorWell} from "src/components/error-well"
 import {errorToString} from "src/util/error-to-string"
 import {call} from "src/util/call"
