+++ apps/zui/src/views/preview-load-modal/shaper.tsx
@@ -2,7 +2,7 @@ import {IconButton} from "src/components/icon-button"
 import styles from "./shaper.module.css"
 import {ZedEditor, useZedEditorKeyboardSubmit} from "src/components/zed-editor"
 import DragAnchor from "src/components/drag-anchor"
-import useSelect from "src/app/core/hooks/use-select"
+import useSelect from "src/util/hooks/use-select"
 import {useDispatch} from "src/app/core/state"
 import {memo, useRef} from "react"
 import LoadDataForm from "src/js/state/LoadDataForm"
