+++ apps/zui/src/views/results-pane/inspector.tsx
@@ -7,7 +7,7 @@ import {useDispatch} from "src/app/core/state"
 import {viewLogDetail} from "src/js/flows/viewLogDetail"
 import * as zed from "@brimdata/zed-js"
 import {valueContextMenu} from "src/app/menus/value-context-menu"
-import useSelect from "src/app/core/hooks/use-select"
+import useSelect from "src/util/hooks/use-select"
 import {ListViewApi} from "src/zui-kit"
 import {PathView} from "./path-view"
 import {AlertView} from "./alert-view"
