+++ apps/zui/src/views/right-pane/history/use-entry-menu.ts
@@ -1,5 +1,5 @@
 import {useDispatch} from "react-redux"
-import useSelect from "src/app/core/hooks/use-select"
+import useSelect from "src/util/hooks/use-select"
 import {showContextMenu} from "src/core/menu"
 import Current from "src/js/state/Current"
 import SessionHistories from "src/js/state/SessionHistories"
