+++ apps/zui/src/views/session-page/editor-resizer.tsx
@@ -2,7 +2,7 @@ import DragAnchor from "src/components/drag-anchor"
 import {useDispatch} from "../../app/core/state"
 import Layout from "src/js/state/Layout"
 import {useRef} from "react"
-import useSelect from "../../app/core/hooks/use-select"
+import useSelect from "../../util/hooks/use-select"
 
 export function EditorResizer({container}) {
   const dispatch = useDispatch()
