+++ apps/zui/src/views/session-page/editor.tsx
@@ -3,9 +3,9 @@ import React, {useRef} from "react"
 import {useSelector} from "react-redux"
 import MainEditor from "src/js/state/Editor"
 import {useDispatch} from "src/app/core/state"
-import {cmdOrCtrl} from "src/app/core/utils/keyboard"
+import {cmdOrCtrl} from "src/util/keyboard"
 import Config from "src/js/state/Config"
-import {useTabId} from "src/app/core/hooks/use-tab-id"
+import {useTabId} from "src/util/hooks/use-tab-id"
 import {ZedEditor} from "src/components/zed-editor"
 import {EditorResizer} from "src/views/session-page/editor-resizer"
 import styles from "./editor.module.css"
