+++ apps/zui/src/views/session-page/pins/base-pin.tsx
@@ -5,7 +5,7 @@ import {useSelector} from "react-redux"
 import {useDispatch} from "src/app/core/state"
 import Editor from "src/js/state/Editor"
 import classNames from "classnames"
-import mergeRefs from "src/app/core/utils/merge-refs"
+import mergeRefs from "src/util/merge-refs"
 import usePinDnd from "./use-pin-dnd"
 import useCallbackRef from "src/js/components/hooks/useCallbackRef"
 import {QueryPin} from "src/js/state/Editor/types"
