+++ apps/zui/src/views/session-page/toolbar.tsx
@@ -12,7 +12,7 @@ import {editQuery} from "src/domain/session/handlers"
 import Layout from "src/js/state/Layout"
 import classNames from "classnames"
 import {useTitleForm} from "./use-title-form"
-import useSelect from "src/app/core/hooks/use-select"
+import useSelect from "src/util/hooks/use-select"
 import {ButtonMenu} from "src/components/button-menu"
 import {useMenuExtension} from "src/core/menu"
 import {createSelector} from "reselect"
