+++ apps/zui/src/zui-kit/react/list-view.tsx
@@ -16,7 +16,7 @@ import {ReactAdapterProps} from "./types"
 import {useStateControllers} from "./use-state-controllers"
 import {mergeRefs, useInitialScrollPosition, useOnScroll} from "./utils"
 import classNames from "classnames"
-import {useParentSize} from "src/app/core/hooks/use-parent-size"
+import {useParentSize} from "src/util/hooks/use-parent-size"
 import {
   TopShadow,
   useScrollShadow,
